add a default http return code to components.
The default http return code is by default 200 HTTP_OK but may be configured per
mapping.  The special value DECLINED instructs tntnet to continue with the next
mapping.

The ecpp compiler generates now a return code DEFAULT, which instructs tntnet to
fall back to that dfeault http return code.

The mvc templates use that feature so that it is not more necessary to return
DECLINED in controllers but just fall throug the end of the component.