change the way arrays are received.
The <%args> section defines the query parameters. When parameters
are defined with a [] at the end of the name, they were treated
as arrays. Multiple values with the same key may be sent and are put into
a std::vector.

No I changed it, that the [] is included in the parameter name. So a arg
definition a[] collects all values with the name "a[]". Previously it
was just "a".

This will break compatibility with existing application but I hope that there
are not too many applications affected.

The reason for the change is, that common javascript libraries work that way.
No e.g. in jquery you can write something like:
  $('#mycontent').load("/page.html", { a = [ 3, 4, 7 ] })

In the ecpp you can then define a array "a[]" and you get those values.