Load ssl certificate file to listener instead of into each request
Cxxtools is now improved to allow loading the ssl certificate to the server
socket and pass it to the clients automatically. This leads to improved perfomance
(+56% on ssl handshake). Also the certificate is loaded at startup and not on
normal runtime any more. When tntnet changes the user, the certificate is already
loaded and hence the new user do not need read access to the certificate and
the key.

Cxxtools also uses SSL_CTX_use_certificate_chain_file to load the certificate
including the certificate chain. Previously SSL_use_certificate_file was used,
which does not load certificate chains.